requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
dominate==2.6.0
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
python-decouple==3.4
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.6
visitor==0.1.3
Werkzeug==1.0.1
